# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TestPilot AI API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/reports/summary": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Summary",
        "operationId": "get_summary_api_v1_projects__project_id__reports_summary_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/reports/issues-by-severity": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Issues By Severity",
        "operationId": "get_issues_by_severity_api_v1_projects__project_id__reports_issues_by_severity_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuesBySeverityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/reports/run-history": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Run History",
        "operationId": "get_run_history_api_v1_projects__project_id__reports_run_history_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "1-indexed page number.",
              "default": 1,
              "title": "Page"
            },
            "description": "1-indexed page number."
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page (max 100).",
              "default": 25,
              "title": "Page Size"
            },
            "description": "Items per page (max 100)."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/reports/trend": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Trend",
        "description": "Future scope (FR-108) \u2014 pass-rate trend line over time.",
        "operationId": "get_trend_api_v1_projects__project_id__reports_trend_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/current/reports/rollup": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Organization Rollup",
        "description": "Future scope (FR-111) \u2014 Organization-level rollup across all projects.",
        "operationId": "get_organization_rollup_api_v1_organizations_current_reports_rollup_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IssuesBySeverityResponse": {
        "properties": {
          "minor": {
            "type": "integer",
            "title": "Minor"
          },
          "major": {
            "type": "integer",
            "title": "Major"
          },
          "critical": {
            "type": "integer",
            "title": "Critical"
          },
          "blocker": {
            "type": "integer",
            "title": "Blocker"
          }
        },
        "type": "object",
        "required": [
          "minor",
          "major",
          "critical",
          "blocker"
        ],
        "title": "IssuesBySeverityResponse"
      },
...
      "ReportSummaryResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "passed": {
            "type": "integer",
            "title": "Passed"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "pass_percentage": {
            "type": "number",
            "title": "Pass Percentage"
          },
          "failure_percentage": {
            "type": "number",
            "title": "Failure Percentage"
          },
          "coverage_percentage": {
            "type": "number",
            "title": "Coverage Percentage"
          }
        },
        "type": "object",
        "required": [
          "total",
          "passed",
          "failed",
          "skipped",
          "pass_percentage",
          "failure_percentage",
          "coverage_percentage"
        ],
        "title": "ReportSummaryResponse"
      },
...
      "RunHistoryResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TestRunPublic"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "RunHistoryResponse"
      },
...
      "TestRunPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "status": {
            "$ref": "#/components/schemas/TestRunStatus"
          },
          "summary_total": {
            "type": "integer",
            "title": "Summary Total"
          },
          "summary_passed": {
            "type": "integer",
            "title": "Summary Passed"
          },
          "summary_failed": {
            "type": "integer",
            "title": "Summary Failed"
          },
          "summary_skipped": {
            "type": "integer",
            "title": "Summary Skipped"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "status",
          "summary_total",
          "summary_passed",
          "summary_failed",
          "summary_skipped",
          "started_at",
          "completed_at",
          "created_at"
        ],
        "title": "TestRunPublic"
      },
...
      "TestRunStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "completed",
          "error"
        ],
        "title": "TestRunStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}